Move image processing code (to improve logging)
